# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчёта трудовых норм и учёта производственных заказов",
        "title": "Сервис нормирования API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/all_final_order": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить данные финального отчёта ПЭО",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала периода (YYYY-MM-DD), по умолчанию — начало текущего месяца",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода (YYYY-MM-DD), по умолчанию — конец текущего месяца",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Номер заказа",
                        "name": "order_num",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Типы изделий (можно указать несколько)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get.FinalReportResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат даты 'from' или 'to'",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/nashchelnik/raw/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить данные по нащельникам",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа (legacy)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.NashchelnikRawData"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/orders/order-norm/by-order": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить нормированные заказы по номеру",
                "parameters": [
                    {
                        "type": "string",
                        "description": "номер нормированного заказа",
                        "name": "order_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "позиция нормированного заказа",
                        "name": "position",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GetOrderDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid position",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/order-norm/{id}/details": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить детали заказа с учётом источника",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Источник данных: mosquito, vodootliv (по умолчанию — обычный заказ)",
                        "name": "source",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GetOrderDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "REQUIRES_CALCULATOR",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/orders/order/norm/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить нормированные заказы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "номер заказа",
                        "name": "order_num",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "тип заказа",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GetOrderDetails"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/order/norm/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить нормированный заказ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID нормированного заказа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.GetOrderDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Нормировка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/vitr-assign": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Получить данные по витражам",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GetWorkersVitrage"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "get.FinalReportResponse": {
            "type": "object",
            "properties": {
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.GetWorkers"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.PEOProduct"
                    }
                }
            }
        },
...
        "storage.AssignedWorker": {
            "type": "object",
            "properties": {
                "actual_minutes": {
                    "type": "number"
                },
                "actual_value": {
                    "type": "number"
                },
                "employee_id": {
                    "type": "integer"
                }
            }
        },
...
        "storage.GetOrderDetails": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "head_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "order_num": {
                    "type": "string"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "parent_product_id": {
                    "type": "integer"
                },
                "part_type": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "profile": {
                    "type": "string"
                },
                "ready_date": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "systema": {
                    "type": "string"
                },
                "template_code": {
                    "type": "string"
                },
                "total_time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.GetWorkers": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "storage.GetWorkersVitrage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "storage.NashchelnikRawData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "customer": {
                    "type": "string"
                },
                "existing_ops": {
                    "description": "Только \"Разметка\", \"Резка\" и т.д.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "legacy_id": {
                    "type": "integer"
                },
                "order_num": {
                    "type": "string"
                },
                "pgm": {
                    "type": "number"
                },
                "sqr": {
                    "type": "number"
                }
            }
        },
        "storage.NormOperation": {
            "type": "object",
            "properties": {
                "assign_workers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.AssignedWorker"
                    }
                },
                "count": {
                    "type": "number"
                },
                "minutes": {
                    "type": "number"
                },
                "operation_label": {
                    "type": "string"
                },
                "operation_name": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "storage.PEOProduct": {
            "type": "object",
            "properties": {
                "brigade": {
                    "type": "string"
                },
                "coefficient": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "customer_type": {
                    "type": "string"
                },
                "employee_minutes": {
                    "description": "Мапа: employee_id → суммарные минуты",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "employee_value": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "norm_money": {
                    "type": "number"
                },
                "order_num": {
                    "type": "string"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "parent_product_id": {
                    "type": "integer"
                },
                "part_type": {
                    "type": "string"
                },
                "position": {
                    "type": "number"
                },
                "profile": {
                    "type": "string"
                },
                "ready_date": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                },
                "sqr_stv": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "systema": {
                    "type": "string"
                },
                "total_time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                }
            }
        },
...
    }
}